refactor(parser): split lexer into multiple files
