+++ .typos.toml
@@ -8,7 +8,7 @@ extend-exclude = [
   "tasks/coverage/babel",
   "tasks/coverage/typescript",
   "tasks/prettier_conformance/prettier",
-  "crates/oxc_parser/src/lexer/mod.rs",
+  "crates/oxc_parser/src/lexer/byte_handlers.rs",
   "crates/oxc_linter/fixtures",
   "crates/oxc_linter/src/rules/jsx_a11y/img_redundant_alt.rs",
   "crates/oxc_syntax/src/xml_entities.rs",
